+++ frontend/src/metabase/nav/components/search/RecentsList/RecentsListContent.tsx
@@ -80,7 +80,7 @@ export const RecentsListContent = ({
                   <ResultTitle
                     data-testid="recently-viewed-item-title"
                     truncate
-                    href={!onClick ? getItemUrl(item) : undefined}
+                    href={onClick ? undefined: getItemUrl(item)}
                   >
                     {getItemName(item)}
                   </ResultTitle>
