+++ frontend/src/metabase/search/components/InfoText/InfoTextEditedInfo.styled.tsx
